Consolidate state package
